+++ lib/flutter_feedback_kit.dart
@@ -1,4 +1,7 @@
-library flutter_feedback_kit;
 
 export 'src/domain/entities/feedback_category.dart';
 export 'src/domain/entities/feedback_entry.dart';
+export 'src/domain/repositories/feedback_backend.dart';
+export 'src/data/backends/webhook_backend.dart';
+export 'src/presentation/feedback_widget.dart';
+export 'src/presentation/feedback_button.dart';