+++ lib/src/domain/repositories/feedback_backend.dart
@@ -1,0 +1,5 @@
+import '../entities/feedback_entry.dart';
+
+abstract class FeedbackBackend {
+  Future<void> submit(FeedbackEntry entry);
+}